// src/enums/router.enum.ts
export enum ERouter {
  VideoIntro = 'VideoIntro',
  PropertyVideo = 'PropertyVideo',
  AuthCallback = 'AuthCallback',
  AuthVerify = 'AuthVerify',
  RedirectLogin = 'RedirectLogin',
  Home = 'Home',
  DeepLinkLanding = 'DeepLinkLanding',
  Chatbot = 'Chatbot'
}

export enum EPropertyRouter {
  PropertyLanding = 'PropertyLanding'
}